requirements.txt
mcp[cli]
biopython
requests
chromadb
sentence-transformers
networkx
streamlit
pandas
plotly
python-dotenv
streamlit-agraph